split plated and non-plated models in test_valid_models
